refactor: split vote-validation.pipe into multiple pipes - add one for create, update, and delete
