+++ src/votes/votes.resolver.ts
@@ -19,7 +19,9 @@ import { CreateVotePayload } from "./models/create-vote.payload";
 import { UpdateVoteInput } from "./models/update-vote.input";
 import { UpdateVotePayload } from "./models/update-vote.payload";
 import { Vote } from "./models/vote.model";
-import { VoteValidationPipe } from "./pipes/vote-validation.pipe";
+import { CreateVoteValidationPipe } from "./pipes/create-vote-validation.pipe";
+import { DeleteVoteValidationPipe } from "./pipes/delete-vote-validation.pipe";
+import { UpdateVoteValidationPipe } from "./pipes/update-vote-validation.pipe";
 import { VotesService } from "./votes.service";
 
 @Resolver(() => Vote)
@@ -53,7 +55,7 @@ export class VotesResolver {
   }
 
   @Mutation(() => CreateVotePayload)
-  @UsePipes(VoteValidationPipe)
+  @UsePipes(CreateVoteValidationPipe)
   async createVote(
     @Args("voteData") voteData: CreateVoteInput,
     @CurrentUser() user: User
@@ -62,13 +64,13 @@ export class VotesResolver {
   }
 
   @Mutation(() => UpdateVotePayload)
-  @UsePipes(VoteValidationPipe)
+  @UsePipes(UpdateVoteValidationPipe)
   async updateVote(@Args("voteData") voteData: UpdateVoteInput) {
     return this.votesService.updateVote(voteData);
   }
 
   @Mutation(() => Boolean)
-  @UsePipes(VoteValidationPipe)
+  @UsePipes(DeleteVoteValidationPipe)
   async deleteVote(@Args("id", { type: () => Int }) id: number) {
     return this.votesService.deleteVote(id);
   }
